backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
exceptiongroup==1.3.0
fastapi==0.116.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
langchain
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.20
sniffio==1.3.1
starlette==0.47.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0

# ML & Embedding Dependencies (Epic 2.2)
torch>=2.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
huggingface-hub>=0.16.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Vector Database (Epic 2.3)
weaviate-client>=3.24.0
